azure/Simgine/requirements.txt
azure-functions
azure-storage-blob
gensim
PyYAML